+++ Commands/Minigames/PestControl.cs
@@ -9,7 +9,7 @@ namespace Supay.Bot {
         return;
       }
       string rsn = string.Empty;
-      int target = 0;
+      long target = 0;
       string skill = string.Empty;
       int points = 0;
       int rank = 0;
@@ -77,7 +77,7 @@ namespace Supay.Bot {
         }
       }
 
-      int curExp = skillInfo.Exp;
+      long curExp = skillInfo.Exp;
       int curLvl = skillInfo.Level;
       if (target == 0 || target <= curExp) {
         target = (skillInfo.VLevel + 1).ToExp();
@@ -94,7 +94,7 @@ namespace Supay.Bot {
       } else {
         N = 35.0;
       }
-      int potExp = curExp;
+      long potExp = curExp;
       var reply = new int[3];
       for (int i = 0; i < 3; i++) {
         double bonus;
